+++ lib/canvas/AnimationService.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 
 import 'package:http/http.dart' as http;
