+++ lib/drawer/body.dart
@@ -1,8 +1,6 @@
-import 'dart:async';
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 import 'package:x/logic/drawerStream.dart';
 
 class Body extends StatefulWidget {
@@ -17,6 +15,7 @@ class Body extends StatefulWidget {
 class _BodyState extends State<Body> {
   @override
   DrawerStream drawerStream = DrawerStream();
+  @override
   Widget build(BuildContext context) {
     //  StreamBuilder(
     //               stream: drawerStream.broadcastStream,
